convert base class into a module
